requirements.txt
# Django
django>=5.0.3,<6.0.0 # 장고 버전이 5.0.1 이상이거나 6.0.0 미만의 버전으로 설치

# DRF (Django Rest Framework)
djangorestframework>=3.14.0,<4.0.0

# Psycopg2
psycopg2>=2.9.9,<3.0.0

# DRF-Spectacular
drf_spectacular>=0.27.1,<0.28.0

# Help Django Chatting Socket Connection
channels>=4.0.0,<4.0.9

# docker-compose build